consumer/requirements.txt
clickhouse-connect==0.8.7
kafka-python-ng[zstd]==2.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
structlog==24.4.0
orjson==3.10.11
